linux-kipr: Add support for the botball-v2 board
This adds a patch adapting the existing am335x-pepper device tree file
to support the Botball-v2.  Additionally, pdata-quirks are added to
support the use of the Wilink8 chipset on this board.  Note that
bluetooth function requires a hciattach to ttyO5 (in userspace).

This commit hasn't been tested within this layer and the kernel patch
itself may yet require further adaptation.

Signed-off-by: Ash Charles <ashcharles@gmail.com>